backend/lambda_minimal/requirements_minimal.txt
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=0.19.0
PyJWT>=2.8.0
tqdm>=4.64.0
chardet>=5.0.0
tiktoken>=0.4.0
faiss-cpu>=1.7.4
numpy>=1.21.0 